// fema-web-client/src/libs/api-client/model/fMStructureDto.ts
/**
 * Fema API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FMStructureDto { 
    code?: string | null;
    longName?: string | null;
    shortName?: string | null;
    category?: string | null;
    level?: number;
    childFMStructures?: Array<FMStructureDto> | null;
}